Rename truncation module to censoring
