+++ aeppl/__init__.py
@@ -11,11 +11,11 @@ from aeppl.printing import latex_pprint, pprint
 
 # isort: off
 # Add rewrites to the DBs
+import aeppl.censoring
 import aeppl.cumsum
 import aeppl.mixture
 import aeppl.scan
 import aeppl.tensor
 import aeppl.transforms
-import aeppl.truncation
 
 # isort: on
